Restructured scripts according to db source
