--- DataForSequelizeAPI/app2.js
@@ -1,58 +1,0 @@
-var fs = require("fs");
-var artists = "RemovedDuplicates2.json";
-var ws = 'works.json';
-var rels = 'releases.json';
-var wsAPI = 'WorksAPI.json';
-var relsAPI = 'ReleasesAPI.json';
-
-var newObj = {};
-var musicians = [];
-var releases = [];
-var work = [];
-
-    // SCRIPT THAT ADDS WORKS AND RELEASES FOREIGN KEYS TO ARTISTS TABLE
-    fs.readFile(relsAPI, 'utf8', function readFileCallback(err, data){
-        if (err) {
-            console.log(err);
-            return;
-        }
-
-        obj = JSON.parse(data);
-
-        fs.readFile(artists, 'utf8', function readFileCallback(err, data2){
-        if (err) {
-            console.log(err);
-            return;
-        }
-
-        ReadArtists = JSON.parse(data2);   
-
-        for (j = 0; j < ReadArtists.length; j++) {
-            ReadArtists[j].release = [];
-        }  
-
-        for (i = 0; i < obj.length; i++) {
-            var releases = [];
-            for (z = 0; z < ReadArtists.length; z++) {
-                if (obj[i].artistId == ReadArtists[z].artistId) {
-                    
-                    var word = obj[i].id.toString();
-                    ReadArtists[z].release.push(word);
-                }
-
-            }
-        }
-
-        json2 = JSON.stringify(ReadArtists); //convert it back to json
-        fs.writeFile('ArtistsAPI.json', json2, 'utf8', function writeFileCallback(err, data) {
-            process.exit();
-        }); // write it back
-          
-          
-        });
-    });
-
-
-
-
-
