+++ musicbrainz/scrapeArtists/.gitignore
