--- scrapeIDArtists/Readme.md
@@ -1,4 +1,0 @@
-Scrapes the IDs' of artists/composers that were born between 1620 and 1820 from Musicbrainz database. The IDs will be used to scrape the works/recordings/releases of the artists. 
-
-"npm install"
-"node server.js" to run
